:art: style: change package name
