--- app/src/main/java/sap/ass01/layered/persistance/MongoDB/MongoEBikeRepository.java
@@ -1,4 +1,0 @@
-package sap.ass01.layered.persistance.MongoDB;
-
-public class MongoEBikeRepository {
-}
