--- app/src/main/java/sap/ass01/layered/persistance/Repositories/UserRepository.java
@@ -1,4 +1,0 @@
-package sap.ass01.layered.persistance.Repositories;
-
-public class UserRepository {
-}
