--- app/src/main/java/sap/ass01/layered/persistance/inMemory/InMemoryRideRepository.java
@@ -1,4 +1,0 @@
-package sap.ass01.layered.persistance.inMemory;
-
-public class InMemoryRideRepository {
-}
