--- app/src/main/java/sap/ass01/layered/persistance/mysql/MySQLUserRepository.java
@@ -1,4 +1,0 @@
-package sap.ass01.layered.persistance.mysql;
-
-public class MySQLUserRepository {
-}
