+++ app/src/main/java/sap/ass01/layered/persistence/MongoDB/MongoRideRepository.java
@@ -1,0 +1,4 @@
+package sap.ass01.layered.persistence.MongoDB;
+
+public class MongoRideRepository {
+}
