+++ app/src/main/java/sap/ass01/layered/persistence/Repositories/EBikeRepository.java
@@ -1,0 +1,4 @@
+package sap.ass01.layered.persistence.Repositories;
+
+public class EBikeRepository {
+}
