+++ app/src/main/java/sap/ass01/layered/persistence/Repositories/UserRepository.java
@@ -1,0 +1,4 @@
+package sap.ass01.layered.persistence.Repositories;
+
+public class UserRepository {
+}
