+++ app/src/main/java/sap/ass01/layered/persistence/inMemory/InMemoryRideRepository.java
@@ -1,0 +1,4 @@
+package sap.ass01.layered.persistence.inMemory;
+
+public class InMemoryRideRepository {
+}
