+++ app/src/main/java/sap/ass01/layered/persistence/mysql/MySQLRideRepository.java
@@ -1,0 +1,4 @@
+package sap.ass01.layered.persistence.mysql;
+
+public class MySQLRideRepository {
+}
