+++ app/src/main/java/sap/ass01/layered/persistence/mysql/MySQLUserRepository.java
@@ -1,0 +1,4 @@
+package sap.ass01.layered.persistence.mysql;
+
+public class MySQLUserRepository {
+}
